+++ FawrySystemService/src/main/java/com/FawrySystem/FawrySystemService/models/payment/CreditCardStrategy.java
@@ -11,7 +11,7 @@ public class CreditCardStrategy implements PaymentStrategy {
     }
 
     @Override
-    public boolean pay(double amount,  Customer customer) {
+    public boolean pay(float amount,  Customer customer) {
         return (creditCard.getCardNumber().length() == 12) && (creditCard.getCvv().length() == 4)
                 && (creditCard.getDateOfExpiry() <= 2030);
     }
