+++ FawrySystemService/src/main/java/com/FawrySystem/FawrySystemService/models/payment/WalletPaymentStrategy.java
@@ -4,8 +4,8 @@ import com.FawrySystem.FawrySystemService.models.Users.Customer;
 
 public class WalletPaymentStrategy implements PaymentStrategy {
 
-    public boolean pay(double amount, Customer customer) {
-        double currentWallet = customer.getWallet();
+    public boolean pay(float amount, Customer customer) {
+        float currentWallet = customer.getWallet();
         if (currentWallet >= amount) {
             customer.setWallet(currentWallet-amount);
             return true;
